refactor(#86): Writer 클래스 이전 - Writer가 하나의 API가 아니라 여러 API에서 등장해서 별도의 파일로 뺌
