+++ app/src/main/java/com/ssafy/neegongnaegong/data/model/studygroup/response/Writer.kt
@@ -1,0 +1,8 @@
+package com.ssafy.neegongnaegong.data.model.studygroup.response
+
+data class Writer(
+    val userId: Long,
+    val name: String,
+    val profileImg: String,
+    val groupRole: String,
+)
